internal/rfb: move splitAnnexB into the rfb package
The kvm-package-level vnc_h264.go and vnc_h264_test.go were the only
tests in package kvm, and that package transitively imports
internal/native — which links libjknative.a, only built for ARM.
Running 'go test ./...' on the CI's x86_64 Linux host therefore
failed at link time as soon as a test file made package kvm
participate in the test build.

Move the H.264 Annex-B splitter (and its tests) into internal/rfb,
where it logically belongs: the encoding-50 path is the only
consumer, and the rfb package has no platform-specific imports so
it tests cleanly on any host. Export it as rfb.SplitAnnexB.

Drop the trivial scaleCoord/rising helpers' tests with this; the
functions are five lines each and exercised end-to-end on the
device.
