+++ internal/rfb/h264_test.go
@@ -1,4 +1,4 @@
-package kvm
+package rfb
 
 import (
 	"bytes"
@@ -11,7 +11,7 @@ func TestSplitAnnexB4ByteStartCode(t *testing.T) {
 		0x00, 0x00, 0x00, 0x01, 0x67, 0x42, 0x00, 0x1E, // SPS (type 7)
 		0x00, 0x00, 0x00, 0x01, 0x68, 0xCE, 0x3C, 0x80, // PPS (type 8)
 	}
-	got := splitAnnexB(frame)
+	got := SplitAnnexB(frame)
 	if len(got) != 2 {
 		t.Fatalf("got %d NALs, want 2", len(got))
 	}
@@ -34,7 +34,7 @@ func TestSplitAnnexB3ByteStartCode(t *testing.T) {
 	frame := []byte{
 		0x00, 0x00, 0x01, 0x65, 0xAA, 0xBB, // 3-byte start code, slice
 	}
-	got := splitAnnexB(frame)
+	got := SplitAnnexB(frame)
 	if len(got) != 1 {
 		t.Fatalf("got %d NALs, want 1", len(got))
 	}
@@ -49,7 +49,7 @@ func TestSplitAnnexBMixedStartCodes(t *testing.T) {
 		0x00, 0x00, 0x01, 0x68, // 3-byte start
 		0x00, 0x00, 0x00, 0x01, 0x65, 0xAA,
 	}
-	got := splitAnnexB(frame)
+	got := SplitAnnexB(frame)
 	if len(got) != 3 {
 		t.Fatalf("got %d NALs, want 3", len(got))
 	}
@@ -57,44 +57,14 @@ func TestSplitAnnexBMixedStartCodes(t *testing.T) {
 
 func TestSplitAnnexBNoStartCode(t *testing.T) {
 	frame := []byte{0x67, 0x42, 0x00}
-	got := splitAnnexB(frame)
+	got := SplitAnnexB(frame)
 	if len(got) != 1 || !bytes.Equal(got[0], frame) {
 		t.Errorf("expected single NAL with whole frame, got %v", got)
 	}
 }
 
 func TestSplitAnnexBEmpty(t *testing.T) {
-	if got := splitAnnexB(nil); len(got) != 0 {
+	if got := SplitAnnexB(nil); len(got) != 0 {
 		t.Errorf("expected no NALs for nil, got %d", len(got))
-	}
-}
-
-func TestScaleCoord(t *testing.T) {
-	cases := []struct {
-		src, srcMax, dstMax, want int
-	}{
-		{0, 1920, 32767, 0},
-		{1919, 1920, 32767, 32767}, // (1920-1) → exactly dstMax
-		{960, 1920, 32767, 32767 * 960 / 1919},
-		{0, 1, 32767, 0},     // edge: srcMax=1
-		{-5, 1920, 32767, 0}, // negative clamps
-	}
-	for _, c := range cases {
-		got := scaleCoord(c.src, c.srcMax, c.dstMax)
-		if got != c.want {
-			t.Errorf("scaleCoord(%d, %d, %d) = %d, want %d", c.src, c.srcMax, c.dstMax, got, c.want)
-		}
-	}
-}
-
-func TestRising(t *testing.T) {
-	if !rising(0b0000, 0b0001, 0b0001) {
-		t.Errorf("expected rising edge")
-	}
-	if rising(0b0001, 0b0000, 0b0001) {
-		t.Errorf("falling edge reported as rising")
-	}
-	if rising(0b0001, 0b0001, 0b0001) {
-		t.Errorf("steady-state reported as rising")
 	}
 }
